refactor: rename module to utilities
