+++ packages/utilities/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "disable-2fa-with-seedphrase",
+  "name": "@near-wallet/utilities",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
